Directories reorganized, refactoring WIP
