+++ app/pages/Gifts/BuildGiftForm.view/index.ts
@@ -1,0 +1,4 @@
+import { BuildGiftForm } from './BuildGiftForm';
+
+export default BuildGiftForm;
+export { BuildGiftForm };
