+++ app/pages/Gifts/BuildGiftPanel.view/index.ts
@@ -1,0 +1,4 @@
+import { BuildGiftPanel } from './BuildGiftPanel';
+
+export default BuildGiftPanel;
+export { BuildGiftPanel };
