+++ app/pages/Gifts/ConsumeGiftForm.view/index.ts
@@ -1,0 +1,4 @@
+import { ConsumeGiftForm } from './ConsumeGiftForm';
+
+export default ConsumeGiftForm;
+export { ConsumeGiftForm };
