+++ app/pages/Gifts/ConsumeGiftPanel.view/index.ts
@@ -1,0 +1,4 @@
+import { ConsumeGiftPanel } from './ConsumeGiftPanel';
+
+export default ConsumeGiftPanel;
+export { ConsumeGiftPanel };
