+++ app/pages/Gifts/GiftsPage.presenter/index.ts
@@ -1,0 +1,4 @@
+import { GiftsPage } from './GiftsPage';
+
+export default GiftsPage;
+export { GiftsPage };
