+++ app/pages/Gifts/index.ts
@@ -1,0 +1,4 @@
+import { GiftsPage } from './GiftsPage.presenter';
+
+export default GiftsPage;
+export { GiftsPage };
