+++ app/routes.tsx
@@ -7,7 +7,7 @@ import UnlockWalletGuard from './components/UnlockWalletGuard';
 import WalletGuard from './components/WalletGuard';
 import routePaths from './constants/routePaths';
 import DashboardLayout from './layouts/DashboardLayout';
-import { ContactsPage, HistoryPage, SendReceivePage, SettingsPage } from './pages';
+import { ContactsPage, GiftsPage, HistoryPage, SendReceivePage, SettingsPage } from './pages';
 import { CreateAccountView, ImportAccountView, UnlockWalletView } from './views/auth';
 import NotFoundView from './views/errors/NotFoundView';
 import {
@@ -15,7 +15,6 @@ import {
   ChangePinView,
   ConfigureFullServiceView,
   DashboardView,
-  GiftingView,
   RetrieveEntropyView,
   PrivacyPolicyView,
   TermsOfUseView,
@@ -102,7 +101,7 @@ const routes: Routes = [
         path: routePaths.APP_TRANSACTION,
       },
       {
-        Component: GiftingView,
+        Component: GiftsPage,
         exact: true,
         path: routePaths.APP_GIFTING,
       },
