--- app/views/wallet/GiftingView/BuildGiftPanel/index.ts
@@ -1,3 +1,0 @@
-import BuildGiftPanel from './BuildGiftPanel';
-
-export default BuildGiftPanel;
