--- app/views/wallet/GiftingView/ConsumeGiftPanel/index.ts
@@ -1,3 +1,0 @@
-import ConsumeGiftPanel from './ConsumeGiftPanel';
-
-export default ConsumeGiftPanel;
