--- app/views/wallet/GiftingView/index.ts
@@ -1,3 +1,0 @@
-import GiftingView from './GiftingView';
-
-export default GiftingView;
