+++ app/views/wallet/index.ts
@@ -3,7 +3,6 @@ export { default as ChangePinView } from './ChangePinView';
 export { default as ConfigureFullServiceView } from './ConfigureFullServiceView';
 export { default as ContactsBookView } from './ContactsBookView';
 export { default as DashboardView } from './DashboardView';
-export { default as GiftingView } from './GiftingView';
 export { default as PrivacyPolicyView } from './PrivacyPolicyView';
 export { default as RetrieveEntropyView } from './RetrieveEntropyView';
 export { default as SettingsView } from './SettingsView';
